Keep viewport fixture test colocated
